+++ flatlaf-core/src/main/java/com/formdev/flatlaf/UIDefaultsLoader.java
@@ -205,20 +205,37 @@ class UIDefaultsLoader
 					if( classLoader != null && !addonClassLoaders.contains( classLoader ) )
 						addonClassLoaders.add( classLoader );
 
-					packageName = packageName.replace( '.', '/' );
 					if( classLoader == null )
 						classLoader = FlatLaf.class.getClassLoader();
 
+					// Get package URL using ClassLoader.getResource(...) because this works
+					// also in named Java modules, even without opening the package in module-info.java.
+					// This extra step is necessary because ClassLoader.getResource("<package>/<file>.properties")
+					// does not work for named Java modules.
+					URL url = classLoader.getResource( packageName.replace( '.', '/' ) );
+					if( url == null ) {
+						LoggingFacade.INSTANCE.logSevere( "FlatLaf: Failed to find package '"
+							+ packageName + "' to load properties files.", null );
+						continue;
+					}
+					String packageUrl = url.toExternalForm();
+					if( !packageUrl.endsWith( "/" ) )
+						packageUrl = packageUrl.concat( "/" );
+
 					for( Class<?> lafClass : lafClasses ) {
-						String propertiesName = packageName + '/' + simpleClassName( lafClass ) + ".properties";
-						try( InputStream in = classLoader.getResourceAsStream( propertiesName ) ) {
-							if( in != null )
-								properties.load( in );
+						URL propertiesUrl = new URL( packageUrl + simpleClassName( lafClass ) + ".properties" );
+
+						try( InputStream in = propertiesUrl.openStream() ) {
+							properties.load( in );
+						} catch( FileNotFoundException ex ) {
+							// ignore
 						}
 					}
 				} else if( source instanceof URL ) {
 					// load from package URL
-					URL packageUrl = (URL) source;
+					String packageUrl = ((URL)source).toExternalForm();
+					if( !packageUrl.endsWith( "/" ) )
+						packageUrl = packageUrl.concat( "/" );
 					for( Class<?> lafClass : lafClasses ) {
 						URL propertiesUrl = new URL( packageUrl + simpleClassName( lafClass ) + ".properties" );